Extract ColorLegendBin into separate file
