+++ charts/MapTab.tsx
@@ -14,7 +14,7 @@ import { ColorLegend, ColorLegendView } from "./ColorLegend"
 import { getRelativeMouse } from "./Util"
 import { ChartConfig } from "./ChartConfig"
 import { MapConfig } from "./MapConfig"
-import { ColorLegendBin } from "./MapData"
+import { ColorLegendBin } from "./ColorLegendBin"
 import { MapProjection } from "./MapProjection"
 import { select } from "d3-selection"
 import { easeCubic } from "d3-ease"
